Refactor prepopulation, (code clean up)
