--- EnTier/Prepopulation/Contracts/IPrepopulationSeed.cs
@@ -1,13 +1,0 @@
-// using Acidmanic.Utilities.Results;
-//
-// namespace EnTier.Prepopulation.Contracts
-// {
-//     public interface IPrepopulationSeed
-//     {
-//
-//         Result Seed();
-//
-//         void Clear();
-//         
-//     }
-// }
