--- EnTier/Prepopulation/LiteContainer.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace EnTier.Prepopulation
-{
-    [Obsolete("This class has been moved to DependencyInjection.LiteContainer.")]
-    public class LiteContainer:DependencyInjection.LiteContainer
-    {
-        
-    }
-}
